requirements.txt
adrf==0.1.9
asgiref==3.8.1
async-property==0.2.2
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.2
django-autoslug==1.9.9
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
google-auth==2.40.3
h11==0.16.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
sqlparse==0.5.3
typing_extensions==4.14.0
update-reqs==1.2.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
